Reorganize project. Split code into multiple files
