+++ lib/air-conditioner-api.js
@@ -15,36 +15,6 @@ function AirConditionerApi(ipAddress, duid, token, log) {
     this.log = log;
 };
 
-const ACFun = {
-    Power: 'AC_FUN_POWER',
-    TempNow: 'AC_FUN_TEMPNOW',
-    TempSet: 'AC_FUN_TEMPSET',
-    OpMode: 'AC_FUN_OPMODE',
-    Direction: 'AC_FUN_DIRECTION',
-    WindLevel: 'AC_FUN_WINDLEVEL'
-};
-
-const OpMode = {
-    Cool: 'Cool',
-    Heat: 'Heat',
-    Wind: 'Wind',
-    Dry: 'Dry',
-    Auto: 'Auto'
-};
-
-const Direction = {
-    SwingUpDown: 'SwingUD',
-    Fixed: 'Fixed'
-};
-
-const WindLevel = {
-    Auto: 'Auto',
-    Low: 'Low',
-    Mid: 'Mid',
-    High: 'High',
-    Turbo: 'Turbo'
-};
-
 AirConditionerApi.prototype = {
     connect: function () {
         this.log('Connecting...');
@@ -52,7 +22,7 @@ AirConditionerApi.prototype = {
         this.controlCallbacks = {};
         this.stateCallbacks = {};
 
-        const pfxPath = path.join(__dirname, 'ac14k_m.pfx')
+        const pfxPath = path.join(__dirname, '../res/ac14k_m.pfx')
 
         const options = {
             pfx: fs.readFileSync(pfxPath),
@@ -170,10 +140,4 @@ AirConditionerApi.prototype = {
 
 util.inherits(AirConditionerApi, events.EventEmitter);
 
-module.exports = {
-    AirConditionerApi: AirConditionerApi,
-    ACFun: ACFun,
-    OpMode: OpMode,
-    Direction: Direction,
-    WindLevel: WindLevel
-}
+module.exports = AirConditionerApi;
