+++ lib/direction.js
@@ -1,0 +1,4 @@
+module.exports = {
+    SwingUpDown: 'SwingUD',
+    Fixed: 'Fixed'
+};
