+++ lib/op-mode.js
@@ -1,0 +1,7 @@
+module.exports = {
+    Cool: 'Cool',
+    Heat: 'Heat',
+    Wind: 'Wind',
+    Dry: 'Dry',
+    Auto: 'Auto'
+};
