+++ lib/state.js
@@ -1,0 +1,8 @@
+module.exports = {
+    Power: 'AC_FUN_POWER',
+    TempNow: 'AC_FUN_TEMPNOW',
+    TempSet: 'AC_FUN_TEMPSET',
+    OpMode: 'AC_FUN_OPMODE',
+    Direction: 'AC_FUN_DIRECTION',
+    WindLevel: 'AC_FUN_WINDLEVEL'
+};
