+++ lib/wind-level.js
@@ -1,0 +1,7 @@
+module.exports = {
+    Auto: 'Auto',
+    Low: 'Low',
+    Mid: 'Mid',
+    High: 'High',
+    Turbo: 'Turbo'
+};
